FPGA_Python/requirements.txt
paho-mqtt
pyserial
typing